Add PapaParse support and spreadsheet upload endpoint
This commit introduces the PapaParse library to handle CSV parsing
within the new spreadsheet module. A new endpoint for uploading
spreadsheets has been added to the router, along with the necessary
controller and utility functions to process the uploaded files. The
package.json and yarn.lock files have been updated to include the new
dependency. Additionally, middleware for the endpoint has been set up
to facilitate the processing of the incoming files.