refactor spreadSheed function to use request body
The spreadSheed function in the controller has been refactored to accept data directly from the request body instead of reading from a file path. This change simplifies the data handling process and removes unnecessary console logs. Additionally, new validation logic has been introduced to ensure the integrity of the incoming data, with the creation of a SpreadSheet type and an initialValidation function. The parsing of CSV content with PapaParse has been commented out for now, indicating a potential shift in how data will be processed in the future. The commit also includes the addition of new files for types and validation, setting up a more robust and maintainable structure for the spread-sheet module.