refact: Update tictHeader to use generic type for header values
This commit updates the `tictHeader` object in the `spread-sheet/shared/headers/tictHeader.ts` file to use a generic type for the header values. Previously, the object was using a specific type (`TictHeaderValues`) for the header values, but now it uses a generic type (`HeadersValues`) that can be used for any header values. This change improves code reusability and flexibility.