refact: Update kiwifyHeader to use generic type for email header
This commit updates the `kiwifyHeader` object in the `spread-sheet/shared/headers/kiwifyHeader.ts` file to use a generic type for the email header. Previously, the header key was 'Cliente Email' and the value was 'buyer_email', but now the header key is 'Email' and the value remains the same. This change improves code reusability and flexibility.