feat: Update processCsvInBackground to use axios for POST request
This commit updates the `processCsvInBackground` function in the `spread-sheet/processCsvInBackground.ts` file to use the `axios` library for making the POST request. The commented out code block that was using `axios.post` has been uncommented, allowing the function to send the `csvChunk` data to the specified URL. This change improves the functionality of the code by enabling the actual sending of the CSV data.