refact: Update hotmartHeader to use generic types for header values
This commit updates the `hotmartHeader` object in the `spread-sheet/shared/headers/hotmartHeader.ts` file to use generic types for the header values. Previously, the object was using specific types for each header value, but now it uses a generic type (`HeadersValues<PerfectpayHeaderValues>`) that can be used for any header values. This change improves code reusability and flexibility.